pyproject.toml
[project]
name = "zapdelivery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "fastapi-mcp>=0.3.3",
    "langchain-groq>=0.3.2",
    "mcp-use>=1.2.7",
    "prisma>=0.15.0",
    "pydantic[email]>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]
